Rework shorewall.conf definition so not sensitive to 'bool-ness' of vars, and retain original file structure
I felt the original definition, just writing out lots of key-value variables, was too prone to user error, and also required that shorewall-boolean values were very carefully crafted as double-quoted strings.

I considered implementing a dict of acceptable keys with a type value (bool, int etc) but discarded that in favour of just repeating the content of the distribution file.

The result is not 'pretty' in jinja terms but achieves the goal. The filter 'yesno' is provided to output a boolean as 'Yes' or 'No'.

The output of values could perhaps be cleaned up with a jinja macro. I left the key names uppercase in the jinja file and runtime-folded so the values could be copy-pasted clearly.